+++ src/partials/shopping-list/shopping-main.html
@@ -1,4 +1,3 @@
-
 <div class="sh-list__container sh-list__removeWraper">
   <div class="sh-list__books__support">
     <include src="./partials/side-bar/support.html">support</include>
@@ -33,10 +32,9 @@
         </picture>
       </div>
     </div>
-      <ul class="listWithBoks"></ul>
-        <div class="button__from-tui">
-          <div id="pagination" class="tui-pagination"></div>
-      </div>
+    <ul class="books__list"></ul>
+    <div class="button__from-tui">
+      <div id="pagination" class="tui-pagination"></div>
     </div>
   </div>
 </div>
